LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.lineTo.nonfinite-expected.txt

2d.path.lineTo.nonfinite
lineTo() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS lineTo() with Infinity/NaN is ignored
